requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
exceptiongroup==1.2.2
fastapi==0.112.2
frozenlist==1.4.1
h11==0.14.0
idna==3.8
multidict==6.0.5
pip==24.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
setuptools==68.2.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
wheel==0.41.2
yarl==1.9.4